modified some tables from the database

+++ database/migrations/2024_05_13_124105_create_results_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
 
             $table->unsignedBigInteger('offer_id');
             $table->unsignedBigInteger('freelancer_id');
-            $table->unsignedInteger('client_id');
+            $table->unsignedBigInteger('client_id');
             $table->string('description');
             $table->boolean('status')->nullable();
